third_party/blink/web_tests/wpt_internal/prerender/session-history-subframe-reload.https.html

<!DOCTYPE html>
<title>Test history.length and reloading a subframe</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/session-history-test-util.js"></script>
<body>
  <script>
    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testSubframeReloadInPrerender"),
        "Passed"
      );
    }, "Subframe reload works in prerendered page");
  </script>
</body>
